# packages/contracts/openapi/api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnePic API",
    "version": "0.1.0",
    "summary": "OnePic 单图模板库的首期受控生图 API",
    "description": "当前 phase-one API 表面（docs/full-stack-architecture.md §6、ADR 0001 D-1）：\n健康、上传、预审、生成/取消/删除/历史、收藏/集合与工作区导出。模板目录仍由\n静态 JSON/TXT/WebP 提供。成功响应为 `{ data, meta }`，失败响应为 `{ error }`；\n字段命名 camelCase，时间 UTC ISO-8601，ID 使用 UUID/ULID，分页使用 cursor。\n本契约是生成类型与运行时校验的唯一事实来源；实现和 generated client 漂移由 CI 阻断。\n"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/uploads": {
      "post": {
        "operationId": "createUpload",
        "tags": [
          "media"
        ],
        "summary": "创建短期上传会话（隔离区）",
        "description": "声明字节数与 MIME，服务端生成对象键（客户端路径一律拒绝）。 字节经 PUT /uploads/{uploadId}/bytes 上传，POST /uploads/{uploadId}/confirm 确认后方可用于预审。会话 1 小时过期。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传会话已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfUploadSession"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "413": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "415": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/uploads/{uploadId}/bytes": {
      "put": {
        "operationId": "uploadBytes",
        "tags": [
          "media"
        ],
        "summary": "上传字节到隔离区",
        "description": "原始字节正文（application/octet-stream），长度必须与创建时声明的 declaredBytes 完全一致；仅会话属主可写。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UploadId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "字节已落盘到隔离区",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfUploadBytes"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "410": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/uploads/{uploadId}/confirm": {
      "post": {
        "operationId": "confirmUpload",
        "tags": [
          "media"
        ],
        "summary": "确认上传完成",
        "description": "校验字节完整后把上传标记为已确认并记录实际 sha256；重复确认返回 409。确认后的对象仍需经预审解码校验才进入 ready。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UploadId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传已确认",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfUploadConfirmed"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "409": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "410": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/prechecks": {
      "post": {
        "operationId": "createPrecheck",
        "tags": [
          "precheck"
        ],
        "summary": "生成前预审",
        "description": "校验模板不可变版本存在、输入对象属主/确认状态、真实解码与像素上限、 模型/质量落在能力注册表内；settings 中携带提示词正文一律拒绝 （PROMPT_REWRITE_BLOCKED）。预审 1 小时过期，生成创建时重新核对 同一预审行绑定的模板版本与输入对象。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrecheckCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "预审通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfPrecheckCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "409": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "413": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "415": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/generations": {
      "get": {
        "operationId": "listGenerations",
        "tags": [
          "workspace"
        ],
        "summary": "本人生成历史（cursor 分页）",
        "description": "仅返回本人任务，稳定排序 created_at DESC + id tiebreaker。 cursor 为不透明签名串（服务端 HMAC 签发，篡改/非法一律 400）， 逐页遍历不重不漏；state 过滤可选。limit 上限 50，默认 20。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/HistoryStateFilter"
          },
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "一页历史（meta.nextCursor 存在时还有下一页）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfGenerationList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      },
      "post": {
        "operationId": "createGeneration",
        "tags": [
          "generation"
        ],
        "summary": "幂等创建生成任务",
        "description": "携带 Idempotency-Key；同键同请求指纹返回同一任务，同键不同指纹返回 409 IDEMPOTENCY_CONFLICT。服务端重新读取不可变模板版本并比对 promptSha256，拒绝浏览器正文替换。任务异步执行，恒返回 202， 经 GET /generations/{generationId} 轮询。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "幂等键；浏览器端在一次点击周期内生成并复用。",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "任务已受理（新建或幂等重放）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfGenerationStatus"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "409": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "429": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/generations/{generationId}": {
      "get": {
        "operationId": "getGeneration",
        "tags": [
          "generation"
        ],
        "summary": "查询生成任务状态与结果元数据",
        "description": "对象级鉴权：仅属主可读（越权 403 FORBIDDEN，不存在 404 NOT_FOUND）。 data 字段以 generation-status-response.schema.json 为准；签名下载 地址与建议轮询间隔在 meta 中（结果媒体存活时才有 downloadUrl， 过期不改变历史成功事实）。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/GenerationId"
          }
        ],
        "responses": {
          "200": {
            "description": "任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfGenerationStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      },
      "delete": {
        "operationId": "deleteGeneration",
        "tags": [
          "generation"
        ],
        "summary": "用户删除生成任务的结果媒体（O01）",
        "description": "仅属主可删（越权 403，不存在 404）。任务/attempt/result 元数据与哈希 事实行保留（历史不改写）；结果媒体字节立即物理删除并写入 deletion_manifest（reason=user_delete），既有签名下载地址即刻 410。 created/queued 任务先按 J08 语义取消（释放配额、job 移除）再删； running 任务记录取消请求后删除（CANCEL_NOT_GUARANTEED，不保证免计 费）；outcome_unknown 拒绝（409 GENERATION_STATE_ILLEGAL，只能经 request-ID 对账退出）。输入媒体在无其他任务引用时一并过期删除。 重复删除幂等：返回一致的 deleted=true。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/GenerationId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "删除结果（含删除后的任务状态）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfGenerationDelete"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "409": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/generations/{generationId}/sidecar": {
      "get": {
        "operationId": "getGenerationSidecar",
        "tags": [
          "generation"
        ],
        "summary": "查询生成任务的可追溯 sidecar",
        "description": "authenticated GET；仅属主可读。成功信封只包含模板、输入、提示词、attempt 与结果的哈希和元数据，不包含提示词正文、Provider 凭据或签名材料。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/GenerationId"
          }
        ],
        "responses": {
          "200": {
            "description": "哈希与元数据 sidecar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfGenerationSidecar"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/generations/{generationId}/cancel": {
      "post": {
        "operationId": "cancelGeneration",
        "tags": [
          "generation"
        ],
        "summary": "协作式取消生成任务",
        "description": "J08 诚实语义：queued 立即取消并释放配额（job 从 pending 队列移除）； running 仅记录取消请求（outcome=cancel_requested， code=CANCEL_NOT_GUARANTEED——provider 可能已接受并计费，上游无取消 API）；已终态任务原样回报（already_terminal）；outcome_unknown 拒绝 取消（409 GENERATION_STATE_ILLEGAL，只能经 request-ID 对账或人工 处置退出）。重复取消幂等：同一任务返回一致状态，配额不会重复释放。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/GenerationId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "取消结果（含 already_terminal 与 cancel_requested）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfGenerationCancel"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "409": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "onepic_session",
        "description": "服务端 opaque 会话（OIDC 授权码 + PKCE 登录后签发，B03）。"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "x-onepic-requested-with",
        "in": "header",
        "required": true,
        "description": "CSRF 护栏自定义头，恒为 onepic-fetch（跨站表单无法设置）。",
        "schema": {
          "type": "string",
          "const": "onepic-fetch"
        }
      },
...
      "UploadId": {
        "name": "uploadId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "PageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "每页条数，默认 20，上限 50。",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 50
        }
      },
...
    },
    "responses": {
      "ValidationFailed": {
        "description": "请求未通过契约校验（含未知字段）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      },
      "Unauthenticated": {
        "description": "无有效会话",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      },
      "ApiFailureResponse": {
        "description": "稳定错误码信封（错误码目录见 D06 数据字典）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      }
    },
    "schemas": {
...
      "UploadCreateRequest": {
        "type": "object",
        "description": "字段以 docs/design/backend-schemas/upload-create-request.schema.json 为准。",
        "additionalProperties": false,
        "required": [
          "declaredBytes",
          "declaredMime"
        ],
        "properties": {
          "declaredBytes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20971520
          },
          "declaredMime": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          }
        }
      },
...
      "UploadConfirmRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sha256"
        ],
        "properties": {
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "description": "客户端对已上传字节计算的 SHA-256（hex）。"
          }
        }
      },
...
      "ApiSuccessOfUploadSession": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UploadSession"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiSuccessOfUploadBytes": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UploadBytes"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiSuccessOfUploadConfirmed": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UploadConfirmed"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiSuccessOfPrecheckCreated": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PrecheckCreated"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
    }
  }
}